requirements.txt
flask
python-dotenv
requests
pandas
numpy
ccxt
